Added Database module and DatabaseAdapter class
